+++ native/KotlinIOS/KotlinIOS/TicketTableViewCell.xib
@@ -1,0 +1,56 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="16097" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="16087"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" preservesSuperviewLayoutMargins="YES" selectionStyle="default" indentationWidth="10" rowHeight="70" id="rbq-eA-735" customClass="TicketTableViewCell" customModule="KotlinIOS" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="414" height="70"/>
+            <autoresizingMask key="autoresizingMask"/>
+            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" preservesSuperviewLayoutMargins="YES" insetsLayoutMarginsFromSafeArea="NO" tableViewCell="rbq-eA-735" id="cAT-4S-gbR">
+                <rect key="frame" x="0.0" y="0.0" width="414" height="70"/>
+                <autoresizingMask key="autoresizingMask"/>
+                <subviews>
+                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Passengers" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="ojy-hl-XPb">
+                        <rect key="frame" x="20" y="40" width="89" height="20"/>
+                        <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                        <nil key="textColor"/>
+                        <nil key="highlightedColor"/>
+                    </label>
+                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Price" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="lhJ-el-Le9">
+                        <rect key="frame" x="355" y="24.5" width="39" height="21"/>
+                        <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                        <nil key="textColor"/>
+                        <nil key="highlightedColor"/>
+                    </label>
+                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Ticket" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="zBY-3Z-S1r">
+                        <rect key="frame" x="20" y="10" width="47" height="20"/>
+                        <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                        <nil key="textColor"/>
+                        <nil key="highlightedColor"/>
+                    </label>
+                </subviews>
+                <constraints>
+                    <constraint firstItem="zBY-3Z-S1r" firstAttribute="top" secondItem="cAT-4S-gbR" secondAttribute="top" constant="10" id="Qx9-lm-7DI"/>
+                    <constraint firstItem="zBY-3Z-S1r" firstAttribute="leading" secondItem="cAT-4S-gbR" secondAttribute="leading" constant="20" id="bK3-kg-jjN"/>
+                    <constraint firstItem="lhJ-el-Le9" firstAttribute="centerY" secondItem="cAT-4S-gbR" secondAttribute="centerY" id="cj4-lm-7Sy"/>
+                    <constraint firstItem="ojy-hl-XPb" firstAttribute="top" secondItem="zBY-3Z-S1r" secondAttribute="bottom" constant="10" id="jKR-jc-ygg"/>
+                    <constraint firstAttribute="bottom" secondItem="ojy-hl-XPb" secondAttribute="bottom" constant="10" id="qt9-xT-hqP"/>
+                    <constraint firstItem="ojy-hl-XPb" firstAttribute="leading" secondItem="cAT-4S-gbR" secondAttribute="leading" constant="20" id="ruF-hh-nbZ"/>
+                    <constraint firstAttribute="trailing" secondItem="lhJ-el-Le9" secondAttribute="trailing" constant="20" id="tK3-is-Fqz"/>
+                </constraints>
+            </tableViewCellContentView>
+            <connections>
+                <outlet property="passengersLabel" destination="ojy-hl-XPb" id="2o5-Dr-KbM"/>
+                <outlet property="priceLabel" destination="lhJ-el-Le9" id="KyW-Hb-Mqd"/>
+                <outlet property="ticketLabel" destination="zBY-3Z-S1r" id="3AJ-4F-G0y"/>
+            </connections>
+            <point key="canvasLocation" x="112" y="119"/>
+        </tableViewCell>
+    </objects>
+</document>
